requirements.txt
alembic==0.9.6
bcrypt==3.1.4
blinker==1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
defusedxml==0.5.0
dominate==2.3.1
et-xmlfile==1.0.1
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-OAuthlib==0.9.4
Flask-OpenID==1.2.5
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
oauthlib==2.0.6
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
pycparser==2.18
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
python3-openid==3.1.0
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
SQLAlchemy==1.1.15
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
xlrd==1.1.0
xlwt==1.3.0
